Move NonEmptyTree to core
